Implement explicit timeout for nodes
Currently we consider node as ‘down’ when a service on it has died and
fails to restart. We know that a service has failed by attaching a
monitor to it and reporting to the RC about the failure when process
monitor notification goes through. One of the reasons for failure are
disconnections, when node becomes unreachable. This is OK for most
scenarios but it has been observed that in case of abnormal connection
problems such as a kernel panic on the node, the TCP connection can
linger for a long time and we don't get a timeout in timely manner.
This commit adds an additional mechanism to the existing system: if a
reply doesn't come back from the node (or precisely the monitor on the
node) in a specified timeout, we just notify the RC about the failure
of the monitored services. RC then tries to restart the services as
normal so everything should work the same.