HALON-134: Update SSPL notification about drive failure.
Notification about drive failure should go via logger actuator
in SSPL. This commit changes behaviour to support that way of
notification. So now it's not halon responcibility to update
information in dcs filesystem, but RAS/sspl one.

Changes:
  * New actuator request type was intoduced
  * Halon fixed a way of notification now node where event happens is
    notified, on the contrary to the previous case when everything
    happened on node keeping RC
  * makeLoggerCmd api for creation logger commands was introduced

Current logger commands are not structurized and there is no schema
for them, this will be done in future.