mero-halon: better support for SSPL replies.
When SSPL replies to mero it may have None in UUID field, this
maps into Null in aeson, and lead to a failure during a pattern
matching. This was fixed, by explicitly checking Null case.
Pattern matching is now total so no new exceptions will occur,
all cases where structure is not expected are now logged.

For unstructured additional parsers were added in order to map
replies on the internal halon structures.