HALON-197: Add delay between boot phases.
In debugging HALON-197, it appears that we're seeing a race condition
between service starting in the Mero confd process and mero-mkfs processes
attempting to connect to the confd process. This occurs because the
systemctl call to start Mero returns before Mero starts services.

This issue did not occur before https://github.com/seagate-ssg/halon/pull/665
because the bootstrap process was slightly slower, leaving more time for
confd/rm services to start.

This commit works around the issue by introducing a delay. The 'correct'
solution will be to use Mero service status messages to drive state
transitions, but these are not available yet.