Don't retry requests to replicas after fixed timeouts.
The inconveniencies of retrying with timeouts are discussed in
https://jira.xyratex.com/browse/HALON-44

The group of replicas can be monitored now for disconnections or epoch
changes. In addition, requests may get a notification if the replicas
drop the requests for whatever reason.

EQs, RSs and MM retry requests only if they are notified of problems
with the requests or the replicas.

The clients of EQs, in turn, now have to monitor EQs when sending events.
Retrying is only necessary if the EQs are disconnected or if
notifications of errors in EQs arrive.

The logic in EQs had to be modified so they always send a reply to
clients whether requests are served or not.