halon: Implement polling of the EQs.
This patch fixes a race in EQs. It was possible that while one EQ was
recording the new RC, another EQ was recording events. This events
however would be not sent to the RC for a long time as they are not
noticed by the colocated EQ.

The fix is to poll the event queue state for the RC, instead of
pretending to know at all times the location of the RC. This way any
EQ can record events, and the RC can get them wherever it is.

As a result, the EQ does not compress paths anymore, nor does it forward
events to the EQ colocated with the RC.