halon: Fix race in EQs when recording the RC.
It was possible that while one EQ was recording the new RC, another EQ
was recording events. This events however would be not sent to the RC
for a long time as they are not noticed by the colocated EQ.

The fix is to wait until the RC is recorded before sending pending events
to the RC. This only makes the race less likely, to avoid it completely,
EQs check whether the RC changed while recording an event, and ask the
client to retry if so with the proper node.