State change machinery (#706)
* Extract `ProcessEncode` into its own module.

* Introduce InternalStateChangeEvent.

- Add explicit 'StateCarrier' to stateful objects.
- Extend `HasConfObjState` in terms of `StateCarrier`.
- Add internal state change event.

* Create deferred state changes.

This packages up all the necessary changes
to be made everywhere, in such a way as can
be applied at some subsequent point.

All machinery should be reasonably hidden now :-)

* Move 'createDeferredStateChanges` into conf rules.

Also add (undefined) `cascadeStateSet`, responsible for creating
the full set of state changes from a single AnyStateSet.

* Introduce cascading state changes.

- Write rules in terms of `StateCascadeRule`
- `cascadeStateChange` will evaluate all `StateCascadeRule`s against the object type to generate a set of `AnyStateChange`
- We cycle this until the set of `AnyStateChange` actions remains the same.
- We create `DeferredStateChanges` from this

* Correctly implement createDeferredStateChanges

* Implement `genericApplyDeferredStateChanges`.

This is the primitive function for applying state changes
to the cluster. Atop of this we can build functions that sync
to confd, generate dynamic failure sets etc.

In order to support this, we temporarily introduce
`createSet` for the legacy cases calling `notifyMero`. Ultimately
we should remove all of these cases.

* Add `genericApplyStateChanges`.

* Link SDev state to Disk state.

Again, this is mostly preparatory work, but now we do not need to
systematically update both disk and sdev state - they come from the
same place.

* Fix compilation without USE_MERO

* fixup! Fix compilation without USE_MERO

* fixup! fixup! Fix compilation without USE_MERO

* Remove 'updateDriveState'.

We temporarily maintain `notifyDriveStateChange` until other rules
have been switched to using `InternalObectStateChange`.

* Update code to work with latest master.

Remove uses of updateDriveState.

* Fix new notification scheme.

Add toConfObjState function, that allow to map Halon internal state
on to mero state.