d-p-scheduler: Don't remove monitors when notifications are lost.
If two nodes cannot talk to each other, one node might try to send
a notification to the other when a process dies. The notification
will be lost and the monitor would be removed.

However, the above is wrong, since the other node might still detect
the connection failure and it needs the monitor to produce the
notification.

This patch has monitors added again if notifications of dead processes
are dropped.