Don't set _gitversion.
The .git folder is not always available. In particular it's not there on
our CI slave machines. And in any case using git --describe to identify
builds is problematic because the numeric ordering makes rpm think that
halon-0.11_13_abcde is newer and an upgrade from halon-0.11_12_edcba,
when it might not be (if they come from different branches).