Spring cleanup (#720)
* Remove obsolete cabal.config.

* Remove all traces of pre-Stack Makefile/cabal-install based build system.

It's obsolete. And most certainly bit rotted, because we're not
testing it anymore. So it's effectively dead code.

* Remove run-coverage.sh script.

We should be using `stack build --coverage` for coverage. And where
Stack's coverage support is lacking, contribute fixes upstream. Custom
scripting nearly always a significant burden to maintain and to get
right.