HALON-242: Add metadata pool and pool version.
This adds a metadata pool, distinct from the usual data pool, and creates
a pool version with one disk in each controller.

Other pool versions are created only in non metadata pools.

This does not solve the issue of multiple overlapping (or even
non-overlapping) pools, which may need to be addressed in future.