replicated-log: Fix implementation of withMonitoring.
withMonitoring uses

bracket_ (link pid) (unlink pid) $
  catch ... (\(ProcessLinkException ...) -> ...)

which doesn't work, because the link exception is not caught when it
arrives while unlink has not returned. This patch instead does

catch (bracket_ (link pid) (unlink pid) $ ...)
      (\(ProcessLinkException ...) -> ...)

Which is an improvement, but because unlink is still interruptible,
the link will still be in effect if some other exception arrives when
unlink executes.