replicated-log: Have the ambassador not block.
The ambassador could block on a delay to let replicas elect a leader.
In this interim the ambassador does not answer to any other messages.

We add a timer to the ambassador, so it can continue processing
messages while the timer expires. This required moving timer to the
top level, and augmenting the state of the ambassador loop. Therefore
we created a new record datatype AmbassadorState.

monitorLeaderReplica was also blocking to wait for a reply to a
registry query. This wait has been removed as well.