replicated-log: Don't send old memberships to ambassadors.
When rejecting a recovery request, the replica could send an old
membership to the ambassador. This could cause the ambassador to lose
contact with the group.

In this patch we have replicas send the membership in the recover
request instead. It may not be accurate, but it will allow the
ambassador to talk to replicas which are actually online.