mero-halon: Adopt the new notification interface.
Instead of providing our own halon/side implementation of the notification
interface, we now do calls to the mero library.

Now the notification interface provides the rpc machine that halon uses for
confc and spiel.

Notification mechanism was simplified, because now we need to use callbacks
to see if certain message was delivered or failed to deliver forever.