Skip internal state change test if rmq is not started.
Technically RMQ is not needed for the test itself but the helpers it's
using do require it. We could either switch out the helpers, make the
helpers keep going even if SSPL service fails to start or skip the
test. Currently we skip the test but only temporarily: HALON-282 will
make CI start rabbitmq and no longer skip tests relying on it started.