HALON-291: Fix RG update with incorrect service pid. (#794)
Removed an old rule that allowed receiving out of order service
start messages. This rule is no longer needed because service
start rule is written in a way that it will return service start
logic in case of RC death and message can't come out of order.

In case if it was out of protocol message then this will be caught
by the local monitor.