Introduce SDevState to hold sdev/disk state.
The principal driver for this is to be better able to deal
with transient failures. Previously we would set `TRANSIENT`
and then return to `ONLINE` when the transient failure was
recovered from, but this is incorrect if the disk was previously
in REPAIRED state, for example.

Note that in various places (where we removed `notifyDriveStateChange`)
calls, we have just used `applyStateChanges` rather than
`applyStateChangesCreateFS`, which would have been the more appropriate
function. However, since dynamic pool versions do not work on the Mero
side, and we are pursuing formulaic pool versions to replace them,
there will be no need for this function and using it is pointless.