[HALON-307] Adapt restart logic to new notification
Restart should now generally be much saner. We use a temporary hack to
deal with unwanted notifications from mkfs. Removed the *differentpid
tests: they were desirable in the past because whether pid was present
or not and whether it was was different or not was part of the core
logic to compensate for lack of information. They may be worth
re-adding anyway to test "nothing happens" logic.