HALON-305: Support link reused callback.
Currently halon doesn't use relation between entrypoint request and
link creation callback. But this may be useful in future, so in order
to have complete interaface we provide all interesting infortion to
halon level. On Halon level it's up to Halon wether to use information
or to ignore that, but future improvements will not require changes
in underlying interface.