[HALON-322] Proper M0Node lookup for mero client nodes
Previously we always went through controller but client nodes are not
attached to any controllers. This meant we would not find the relevant
m0node, would not set it to transient and its processes to inhibited
and therefore would not try to restart procesess when the node came
back.

Also minor logging changes.