Make it easier to override Version.hs.
Instead of monkey patching of the version file using sed, which can be
brittle, we instead ask that tools dump a .hs file with all the
version info as they want it. This is used by the Castor build system,
and this method is easier than the alternative.

This patch also replaces the "built on" line with a "authored on" line.
The reason for this is that castor's nix based build procedure resets
the current date to the epoch anyways, so "built on" won't tell you much
at all. Making build artifacts not build on current time furthermore
helps make the build result more deterministic.