replicated-log: Give replicas the initiative to update.
Till now, replicas would only try to get the latest decrees if some
other replica hints to them that they are delayed. This caused the
group to stop when the leader replica was isolated and the followers
were delayed.

To fix this the replicas now propose decrees when there is no known
leader and they notice they are delayed.