HALON-309: Implement keepalive mechanism for SSPL-HL. (#891)
Currently amqp library may silently loose connection to rabbit-mq.
In this case halon will not reply to SSPL-HL requests. In order to
fix this problem we eventually send explicit keepalive messages to
the halon exchange, in this case if HALON does not see keepalive
in TIMEOUT seconds it restarts connection.