CASTOR-1683: Only start using a drive once it's ready. (#892)
* Only start using a drive once it's ready.

To be ready, a device should be present, powered and
visible to the operating system. At this point (and not before)
we may mark it as visible to Mero, attach it, and potentially
start rebalance operations.

We also move from calling `handleRepairInternal` directly and
rely on state transitions and `PoolRebalanceRequest`.

Logic should now be: rules handling SSPL-level failures should
firstly be concerned with updating the state in the graph. Once
this is done, they can call 'checkAndHandleDriveReady' to trigger
any Mero related operations.

* Check that drive is ready before rebalance.

Previously the only thing we cared about was that the drive was
replaced, which did not require that the drive was visible to the
OS. This could lead to starting rebalance to early.

* Add some additional logging.

Also remove an unnecessary line.

* Update power status when drive inserted/removed.

Fix after rebase

Fix tests

* Log why devices might not be ready.

* Use connectUniqueFrom rather than old disconnect style.

* Check status in a better way.

* Silly fix