Don't expect reset confirm if power is off.
First we detect Mero error, and see that the drive isn’t removed and it’s powered.
So we start reset attempt. Then we get note that it’s not powered. Then the reset
attempt starts, sees that it’s not powered, and assumes that it was already powered
down during the reset attempt (and that RC then crashed). Previously it would have
jumped to attempting to power the drive back on. But when we switched to single reset,
it was just updated to wait for an acknoledgement of that reset.
Which never arrives