Allow to check old state when waiting for notification.
This change is an improvement that allow HALON to not
react of meaningless state changes, e.g. drive moved from
Repairing to Failed in case of SNS operation cancelation.

The main impact of this change is to disable drive set LED
commands beign sent for the drive when no repair is happening.