[HALON-437] Improve notification mechanism.
Output more information when notification failed to be delivered
to mero, e.g. output process fid and epoch.

Do not send message to processes that are in undefined state.
Currently undefined state is an initial state, it means that
process is not running, so seding message to such process leads
to round of the delivery-failed notifications.