[HALON-426] Remove double initialization of the hastate in halon:m0d.
Remove double initialization of the hastate, that makes running
halon:m0d finalization much more complex. Also this leads to warnings
in the logs, that are actually expected in this workflow.

So only `withNI` is left that does exactly what previous code was
doing.

This is low risk change, as it doesn't affect other subsystems.