[HALON-426] Allow hastate finalization on halon m0d.
In prior to ha_interface introduction it was not possible to
stop service because it would lead to unloading of the kernel
module. However now it's possible to call ha_state_fini(), that
just stops ha_interface, but does not unload kernel or finalize
ha_interface.

A consequence of this change is that interface will be closed
and status will report correct state for the HA service.