[HALON-441] Handle process restart exit codes
This actually has a wider impact that expected. Since some time
recently, we are actually no longer `start`ing processes but always
`restart`ing them: this is because we set process to starting state
before calling `startMeroProcesses`. It just happened to be OK but it
means that failures would have gone uncaught.

The main change in this PR is that `ruleProcessRestart` now does less:
it no longer explicitly tries to fail services associated with the
process. This rule is most often triggered through a process failed
internal state change notification and if we're receiving that then
cascade would have already failed services as well. This allows us to
use `ruleProcessRestart` as a much more friendly rule to invoke from
other places and to get out process restart timeout and result message
handling that it offers out of it.