HALON-444: Don't mark already FAILED disk as HALON-failed.
HALON-failed is a "protected" state to SSPL - it will not be
overridden. We only set it when we have inferred a failure from
a non-SSPL channel, such as Mero. It was possible for it to be set
on an SSPL-informed failure due to a race with Mero notifications.
This patch updates the check such that it will only be set after
repeated reset attempts.