[HALON-445] Restart rebalance if disk reattached
It may happen that more than one drive fails. For example we detach two
drives, repair happens, rebalance starts. During rebalance, we may
attach one of the drives back and the drive may be just fine. In this
case rebalance might have proceeded "too far" already and skipped over
the drive as it was failed at the time.

To make sure all online drives are taken care of during rebalance, if a
drive is attached during it, just abort and start it again.