[HALON-476] Don't disconnect HALink on keepalive (#1017)
We don't want to manually disconnect ha_links when we detect keepalive:
we want link disconnecting callback to fire first. This will fire when
process in announced as failed but if we disconnect the link then
announce failure, callback never got a chance to run. Now we just leave
the link lying around, announce process as failed and let the callback
deal with actually calling disconnect.