HALON-489: uninhibit processes on node recovery.
When node is recovered from a failure it's processes are marked as
inhibited, this could lead to a various problems, like for offline
processes state will be 'Inhibited Offline' and such process will
be counted as not yet stopped.
As a result new logic was introduced with following rules.

  Inhibited Offline -> Offline
  Inhibited Unknown -> Unknown

All other states will go to failed and lead to a process restart
in case if disposition is ONLINE. This is needed because some
notifications may not be delivered to the process and it has outdated
state.