Implement cardinality in the RG.
This allows specifying at the `Relation` level the desired
number of resources at either end of a relation and using
`G.connect` in place of the various `connectUnique` functions
to correctly specify the relationship. Likewise, `connectedTo`
and `connectedFrom` should now return the appropriate type for
the relationship.

Note that it is not possible to implemeent `ExactlyOne`
semantics.