HALON-501: Make process restart check process comes online.
Previously, the process restart rule terminated as soon as systemd
terminates. If Mero never sends its updated status, then the process
can hang in 'PSStarting' state indefinitely.

This commit now makes the restart job await notification that the
process has come up. If no such notification arrives within 180s,
the process will be marked as "failed to start" and the node will
be failed.

If notification that the process has failed is received before the
process starts, then we check a restart counter, and attempt to
restart it up to 5 times before bailing and failing the node.