[HALON-103] Fix halonctl service message
Original ticket was about "No contact from RC" being reported when
`halonctl … service … start` commands were issued even if the message
got to RC and was processed. In current code, the job handling the
request announces the result in timely manner and in ever case so should
no longer be an issue. This ticket can be closed as-is but I fix the
message shown to the user while here as well as bad flags used for
halon:m0d service (mostly gone unreported because no-one actually starts
it from hctl).