[HALON-522] Verbose for hctl bootstrap cluster
It may happen that facts/roles change after bootstrap happens in which
case it can be very difficult to figure out what halon was actually
started with. `-v` will now print facts and roles used as well as
provide a little more noisy output during start-up as to what it's doing.