[HALON-493] Always wipe drive reset marker
There were multiple cases when `SDOngoingReset` would remain attached to
`StorageDevice` in RG even though the reset rule terminated already. We
now make sure to go through a `finalizer` phase which unsets the
attribute if present (and does nothing if not).

We go through finalizer in the starting body of the rule even though we
may not have set it in that run to account for RC restart:

* reset starts
* drive marked
* RC restarts
* rule starts
* drive may no longer be eligible for reset
* rule ends but marker is still around

Further, on drive attach failure, we actually fail the drive and notify
DM instead of simply doing nothing.

Lastly, remove `ResetSuccess` and `ResetFailure` types and make users
watch for `ResetAttemptResult` instead emitted by the job.