[HALON-373] Implement ruleProcessStart
This rule supersedes restart rule functionality as well as the use of
various helpers which were scattered around the code seemingly randomly.
Now the user only needs to fire-and-forget: dealing with process
configuration, starting, notifications &c. is all performed by the job.

It is still up to the user to ensure the environment is set up properly
however: the rule checks if everything seems OK but will happily return
process start failure if it is not the case. In fact, it will even fail
an already running process if we request start for it but the
environment isn't correct.

Restart process control messages were squashed into start messages which
were unused (due to a bug). Further, start/configure process control
messages are now made to only work over 1 process at a time: this
greatly simplifies logic, removes need for dispatch rules and removes
need for covoluted helpers that built start/configure messages. The
drawback is that we send more messages on the wire: we do not send so
many process messages where this should ever be a problem. We no longer
send fids and endpoints explicitly but instead send proper `Process`
objects. This allows us to retrieve endpoint on halon:m0d side, have
nicer logging, perform nicer comparisons &c. Lastly, results produced by
start/stop/configure control messages inside Either were flipped:
indicate failure in Left and success in Right. This is now consistent
with rest of the system and we don't have awkward side-switching.

HalonVars was extended with multiple new values. This removes any
hard-coded times that were in the code modified in this commit. As it
seems this will be the way to define options at least for short-term and
in light of recent discussion to make these configurable on running
systems through hctl, it makes sense to do so. Also allows us to write
saner tests if we choose to do so.

Any users of the process control channels were modified. Process start
now only happens in 4 places (as far as I could find):

* Bootstrap, boot level 0
* Bootstrap, boot level 1
* Bootstrap, clients
* Process failure -> process start

All of these simply dispatch ruleProcessStart and await results. Due to
HalonVars changes and moving configuration into ruleProcessStart, the
bootstrap procedures do not have timeouts themselves. Previously it
could take a very long time for bootstrap timeout to occur: `mkLoop`
helpers used meant that as long as a message came through sometimes,
many processes could hang for a very long time until failure. Now each
process is responsible for itself and managing its own timeout. This
means that if a process fails or times out, bootstrap procedure will
know sooner and can fail much sooner. This stops scenarios where we had
a 10 minute cluster timeout but it took 30 minutes to actually time out:
whenever a process would reply with result, the timeout refreshed.

We no longer manually set `PSStarting` on service/process in graph.
ruleProcessRestart does a proper notification for process (and for
services through cascading). Mero seems fine with this so this hack is
no longer necessary. We still do this in stop case which was largely
untouched and out of scope of HALON-373.

Tested on devvm and dev1-1. Worked in presence of failure on
dev1-1 (i.e. properly failed bootstrap) due to minor caveat (HALON-537).
Worked through `start -> stop -> start`. `mero-halon:scheduler-tests`
pass. `mero-halon:tests` pass except for expander reset (unrelated to
PR).