[HALON-497] Remove Dynamic failure sets.
Formulaic or optionally preloaded sets should be used.

This lets us throw away a bunch of code used for defining the strategy.
`genders2yaml` is the only remaining user and genders2yaml is no longer
used so it is removed in this commit as well.

Further, we can remove the Strategy middle-man which no longer serves a
purpose and use updates directly.