Don't (re)start process on PSFailed -> PSFailed (#1075)
It may happen that a process in `PSFailed` state changes state to
`PSFailed` as far as state change mechanism is concerned: for example,
reason for failure can change. In this case, we probably don't want to
dispatch another reset.

Alexander has even seen a loop as a result of this: process start job
fails, process is failed, internal state change sees the process
transition, job starts…

This is only a short-term hack: there is nothing stopping more involved
loops. For example it could happen that:

* Process fails
* Start process job starts
* Process put in PSStarting
* Job fails
* applyStateChanges PSFailed
* Job finishes
* Process put in PSFailed
* Job starts…

We could mitigate this by waiting for PSFailed to be announced inside
the job before finishing, therefore increasing the chances of job still
running when `ruleProcessDispatchRestart` fires: but it is still
fragile.

We should probably consider some kind of a `PSUnrecoverable` state,
mapping to `M0_NC_FAILED` that transitions out only after node reboot or
some other conditions.