Fix timeout issue in cep.
CEP didn't update timeout in case if 2 rules were waiting for that.
Assume we have 2 instances of the rules, A and B, A is waiting for
2s, and B is waiting for 4s. When first timeout happens both rules
are awaken, but A continue, and B suspends, however B doesn't renew
timeot, so it's not awaken again.

This commit fixes this situation and now when B is suspended it will
renew 2s timeout.