[HALON-514] Add timeout opt to hctl cluster status
Originally the ticket was about finding out "root cause" of hctl cluster
status not returning. But the ticket makes it clear that it happens
under load. But:

* We already print a message and exit with failure (done recently in
  other issue)

* The code itself is straight forward, single phase: there is not much
  that can be blame on here for failing under load. It was just too
  slow. Custom timeout can just let the user wait for longer.

* Halon itself is now much less resource hungry than it was when the
  issue was created, meaning the situation should be better much more
  often than it was.

Performance improvement and custom timeout should be enough to assure we
get cluster status eventually. If we can't perform a single message
round-trip, we have bigger problems.