[HALON-548] Implement logging RFC in CEP (#1085)
* Rewrite existing CEP logging.

This commit rewrites existing logging in CEP to accord to the
new RFC on logging. It does not yet emit the new log types
defined in that RFC.

Main changes:

Introduction of Application type, which is a bit weird. Basically I didn't want to introduce another type parameter which would be ignored most of the time. Instead, I use the class Application as basically a type-level record. No values of it are expected to exist, it just packages up other types and lets us project to them. This I hope doesn't cause too much churn.

Logs are stripped out of a lot of return types like SM_Complete etc. This is because we now emit logs ASAP, rather than bundling them until the end of a phase.

Most annoying change - we need to carry the local state parameter l along with an SM, because it needs to accord with the local state logger. This would be fine, but it forces the existential up to RuleData and it turns out record syntax with existentials is annoying, so we have to rewrite a bunch of things in Engine.hs to use things like withRuleStack instead.

Factor out `Location` from events.

Capture rule name in phaseLog.

Implement the following log types:

- Fork (some work needed)
- Continue
- Switch
- Stop
- Suspend
- StateLog
- PhaseEntry

Allow setting local state logger.

We now set the logger in `newSM` (via `buildRuleData`), so don't
need to carry it around in `SMExecute`. Which means we can again
remove the existential on `RuleData`, cleaning up a wart from the
initial rewrite.

Revert changes to deal with existential record.

Fix some issues in CEP.

Logger must now be set before definitions.

This is a consequence of the logger now being built into the
SM from the start, rather than carried around in 'SMExecute'.
I think this is a reasonable change.

[skip-ci] Remove forked logging tests from CEP.

Since now logs are emitted as soon as the log command is made,
there is no longer any reason to test their behaviour in forked
environments.

* Update 'halon' to new CEP syntax.

* Transition LoopState->RC.

This commit just contains renames and trivial changes necessary
to use the `Application` class. More semantically relevant changes
will be made in other commits to make review easier.

* Use new logging framework from mero-halon.

Fix mero-halon tests to use new CEP logging.