[HALON-560] Implement resource monitoring service.
The main idea is to be able to get semi-real-time information about GC
statistics. The `ekg` package offers that out of the box. Indeed the
implementation of this part alone was not even the quarter of the code
submitted here.

Here, more of `ekg`'s functionality is bound to. `ekg` exposes various
internal trackers which the user can set. `ekg` will then also report
the values of those. To accomodate these, a serialisable wrapper over
functions on those trackers is provided. A single user-defined metric is
provided here as an example of utility.

Two main issues that I'd like to eventuall address still remain:

* We can't bind to 'modify' functions. Perhaps instead using static
  closures is something to consider. Indeed this may cut down on code
  though I wanted to avoid closures if possible.

* 'set' functions which can't be interleaved will happily leave tracker
  with wrong result if they get re-ordered. For now user should be
  careful about events occuring within a short span of each other and
  races caused by them. Could be partially mitigated with timestamp on
  messages for some use-cases.