[HALON-567] Use SafeCopy for persisted messages
The core changes are small: add constraint to persistent message
creation and actually use it. No changes in CEP required beyond that.

Note that now we should no longer require Binary constraint in a lot of
places. Removal of those is deferred.

It'd be interesting to benchmark the difference between
safecopy-persisted messages and our old way, mainly to measure the
impact of using `cereal` as opposed to `binary` for those.