[HALON-567] Actually make safecopy messages work
We require SafeCopy for creation of persistent messages. This lets GHC
flag up every type that needs an instance.

As CEP only requires Serializable, we apply a trick. For every type we
want persisted, we make the Binary (i.e. Serializable) instance use the
SafeCopy instance. It's achieved by generating the Binary instance when
we're generating SafeCopy instance. As persistance uses SafeCopy
instance directly for storage, this should be OK: it doesn't care what
we do in our Binary instance.

It does mean we have a bit of overhead for persistent messages (because
we're running cereal with binary) but it should be negligable (though
needs measuring). All persistent store values should be as they were
allowing offline upgrade. Online upgrades should be fairly easy (i.e.
write Migrate instances) after this point for in-flight messages.