HALON-570: fix possible possible ignorance of the new drive.
In HALON-570 issued Smart request for the old drive instead
of a new drive, because drive candidate that was created in
HPI rule was never actualized. This happened because HPI
rule incorrectly decided that drive was powered on instead
of beign inserted.

Currently sspl:monitor-status-hpi rule doesn't check if drive
was changed or not, when decide what happen with disk.
That may lead to a problems when halon could make incorrect
decisions. The problematic case is:
0. Drive fails, marked as removed, etc.
1. SSPL-LL sends HPI event for some reason -> this leads to the
   drop of the DriveRemoved status
2. SSPL-LL sends HPI event about new drive, beceuse there is
   no drive removed status, halon decide that it's a repower.

The fix is to make halon aware if it was a new drive or not.
It seems it's not easy to fix `1.` as HPI message is the only
source for drive insertion event, this means that if
monitor-drivemanager and hpi messages will be unordered, then
halon will miss drive insertion.